backend/requirements.txt
# ==== Flask Web Framework ====
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0

# ==== Database & ORM ====
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# ==== Authentication & Security ====
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2

# ==== API & CORS ====
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
flask-restx==1.3.0
Flask-Mail==0.9.1

# ==== Data Validation & Serialization ====
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# ==== Veri ve Görsel İşleme ====
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
opencv-python==4.10.0.84
scikit-image==0.24.0
matplotlib==3.9.2
tqdm==4.67.1

# ==== Derin Öğrenme (PyTorch) ====
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
timm==1.0.11

# ==== TensorFlow (Mac M2 Metal destekli) ====
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
keras==3.10.0

# ==== Yardımcı Kütüphaneler ====
python-dotenv==1.0.0
gunicorn==21.2.0
protobuf==4.25.8
requests==2.32.5
absl-py==2.3.1
rich==14.2.0
h5py==3.14.0

# ==== Testing ====
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# ==== Development ====
python-dateutil==2.8.2
pytz==2023.3
